[refactor]: 파일 분리, 로직 순서 변경, rename
PlayRockScissorPaper, StoreGameData swift 파일 분리
Flow 기준으로 로직 순서 변경
